% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere.R
\name{add_cBot_columns}
\alias{add_cBot_columns}
\title{Helper function: add cBot columns}
\usage{
add_cBot_columns(data)
}
\arguments{
\item{data}{data frame}
}
\value{
data frame with cBot columns
}
\description{
Helper function: add cBot columns
}
\keyword{internal}
