% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_times.R
\name{add_times}
\alias{add_times}
\title{Helper function: add times}
\usage{
add_times(df, sim_datetime_start = as.POSIXct("2017-05-01", tz = "UTC"))
}
\arguments{
\item{df}{df}

\item{sim_datetime_start}{simulation start as datetime object (default:
as.POSIXct("2017-05-01", tz = "UTC"))}
}
\value{
data frame with added  times
}
\description{
Helper function: add times
}
