% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_atmosphere}
\alias{aggregate_atmosphere}
\title{Aggregate Atmosphere}
\usage{
aggregate_atmosphere(atm_selected, format = "wide")
}
\arguments{
\item{atm_selected}{atm_selected}

\item{format}{"wide" or "long}
}
\value{
aggregated data
}
\description{
for hydrologic years
}
