% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_solute}
\alias{aggregate_solute}
\title{Aggregate Solute}
\usage{
aggregate_solute(
  solute,
  sim_datetime_start = as.POSIXct("2017-05-01", tz = "UTC"),
  col_aggr = "date"
)
}
\arguments{
\item{solute}{solute}

\item{sim_datetime_start}{simulation start as datetime object (default:
as.POSIXct("2017-05-01", tz = "UTC"))}

\item{col_aggr}{column to be aggregated. One of "date", "yearmonth", "year" or
"year_hydrologic" (default: "date")}
}
\value{
aggregated solute
}
\description{
Aggregate Solute
}
