% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_tlevel}
\alias{aggregate_tlevel}
\title{Aggregate t_level}
\usage{
aggregate_tlevel(
  t_level,
  sim_datetime_start = as.POSIXct("2017-05-01", tz = "UTC"),
  col_aggr = "date"
)
}
\arguments{
\item{t_level}{t_level}

\item{sim_datetime_start}{simulation start as datetime object (default:
as.POSIXct("2017-05-01", tz = "UTC"))}

\item{col_aggr}{column to be aggregated. One of "date", "yearmonth", "year" or
"year_hydrologic" (default: "date")}
}
\value{
aggregated t_level data
}
\description{
only for columns starting with "sum" and matching "volume"
}
