% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_ctop_with_virtualstorage.R
\name{combine_dataframes}
\alias{combine_dataframes}
\title{Helper function: combine dataframes}
\usage{
combine_dataframes(df_list)
}
\arguments{
\item{df_list}{df_list}
}
\value{
df
}
\description{
Helper function: combine dataframes
}
