% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere.R
\name{copy_column}
\alias{copy_column}
\title{Helper function: copy column}
\usage{
copy_column(df, column, number_of_new_columns)
}
\arguments{
\item{df}{data fram}

\item{column}{name of colum to copy}

\item{number_of_new_columns}{number of columns to be copied}
}
\value{
df with copied column names
}
\description{
Helper function: copy column
}
\examples{
# Beispiel-Tibble
tb <- tibble::tibble(
ID = 1:5,
Wert = c(10, 20, 30, 40, 50)
)

# Anwendung der Funktion
tb_neu <- copy_column(tb, "Wert", 3)
print(tb_neu)

}
