% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hydrus_exe.R
\name{create_hydrus_exe}
\alias{create_hydrus_exe}
\title{Create Hydrus1D executable (only on Windows!)}
\source{
\url{https://github.com/phydrus/source_code/tree/main}
}
\usage{
create_hydrus_exe()
}
\value{
path to Hydrus1D v4.0.8 executable
}
\description{
Create Hydrus1D executable (only on Windows!)
}
