% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere_data.R
\name{create_monthly_atm}
\alias{create_monthly_atm}
\title{Aggregate Atmospheric Data to Monthly Values}
\usage{
create_monthly_atm(atm = prepare_atmosphere_data())
}
\arguments{
\item{atm}{atm as retrieved by \code{prepare_atmosphere_data}}
}
\value{
tibble with yearly atmospheric data values
}
\description{
Aggregate Atmospheric Data to Monthly Values
}
\examples{
create_monthly_atm()

}
