% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydrologic_years.R
\name{get_hydrologic_years}
\alias{get_hydrologic_years}
\title{Get Hydrologic Years}
\usage{
get_hydrologic_years(datetime)
}
\arguments{
\item{datetime}{datetime string}
}
\value{
integer vector with hydrologic year
}
\description{
Get Hydrologic Years
}
