% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservative_tracer-traveltime.R
\name{get_last_day_of_months}
\alias{get_last_day_of_months}
\title{Helper function: get last days of months}
\usage{
get_last_day_of_months(ids, date_start = "2017-05-01")
}
\arguments{
\item{ids}{vector with ids}

\item{date_start}{(default: "2017-05-01")}
}
\description{
Helper function: get last days of months
}
\keyword{internal}
