% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservative_tracer-traveltime.R
\name{get_traveltimes}
\alias{get_traveltimes}
\title{Get traveltimes (for conservative tracers)}
\usage{
get_traveltimes(solute_files, dbg = TRUE)
}
\arguments{
\item{solute_files}{paths to solute files, with good naming convention for
monthly solute exposition (e.g. 0110, solute1: first month, solute10: tenth
month after simulation start)}

\item{print}{debug messages (default: TRUE)}
}
\value{
tibble with time of substance at top/bottom and diff time. Note that
the percentile relate to the substance load
}
\description{
Get traveltimes (for conservative tracers)
}
