% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atmosphere}
\alias{plot_atmosphere}
\title{Plot Atmosphere}
\usage{
plot_atmosphere(atm_selected_hydro_long)
}
\arguments{
\item{atm_selected_hydro_long}{as retrieved by \code{\link{aggregate_atmosphere}}
in "long" format}
}
\value{
plot atmosphere
}
\description{
Plot Atmosphere
}
