% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere_data.R
\name{plot_monthly_atm}
\alias{plot_monthly_atm}
\title{Plot Monthly Atmospheric Data}
\usage{
plot_monthly_atm()
}
\value{
plot of monthly atmospheric data
}
\description{
Plot Monthly Atmospheric Data
}
\examples{
plot_monthly_atm()
}
