% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_solute}
\alias{plot_solute}
\title{Plot Solute}
\usage{
plot_solute(solute_aggr, y_label = "Share of 'clearwater' (\%)")
}
\arguments{
\item{solute_aggr}{solute_aggr as retrieved by \code{\link{aggregate_solute}}}

\item{y_label}{y_label (default: "Share of 'clearwater' (\%)")}
}
\value{
plot solute
}
\description{
Plot Solute
}
