% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservative_tracer-traveltime.R
\name{plot_traveltimes}
\alias{plot_traveltimes}
\title{Plot traveltimes}
\usage{
plot_traveltimes(solute_travel, title = "", ylim = NULL)
}
\arguments{
\item{solute_travel}{tibble as retrieved by \code{get_traveltimes}}

\item{title}{optional title (default: "")}

\item{ylim}{optional ylim (default: NULL)}
}
\value{
plotly of traveltimes for different percentiles
}
\description{
Plot traveltimes
}
