% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_waterbalance}
\alias{plot_waterbalance}
\title{Plot Water Balance}
\usage{
plot_waterbalance(
  tlevel_aggr,
  y_label = "Water Balance Component",
  unit_org = "cm",
  unit_target = "mm"
)
}
\arguments{
\item{tlevel_aggr}{aggregated t_level as retrieved by \code{\link{aggregate_tlevel}}}

\item{y_label}{default: "Water Balance Component (mm)"}

\item{unit_org}{original unit in "t_level" (default: "cm")}

\item{unit_target}{target unit for plot (default: "mm")}
}
\value{
plot water balance
}
\description{
Plot Water Balance
}
