% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_precipitation-daily.R
\docType{data}
\name{precipitation_daily}
\alias{precipitation_daily}
\title{Precipitation: Daily}
\format{
A data.frame with 8835 rows and 2 variables:
\describe{
\item{datetime}{date time}
\item{precipitation_mm}{precipitation in mm}
}
}
\usage{
precipitation_daily
}
\description{
Hourly precipitation data downloaded from DWD for monitoring station Braunschweig
(id = 662) between 1997-10-22 and 2021-12-31, which were aggregated to hourly
values within R
}
\keyword{datasets}
