% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere_data.R
\name{prepare_atmosphere_data}
\alias{prepare_atmosphere_data}
\title{Prepare Atmospheric Data}
\usage{
prepare_atmosphere_data()
}
\value{
data frame with atmospheric data for Braunschweig
}
\description{
Prepare Atmospheric Data
}
\examples{
atm <- prepare_atmosphere_data()
atm

}
