% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_ctop_with_virtualstorage.R
\name{recalculate_ctop_with_virtualstorage}
\alias{recalculate_ctop_with_virtualstorage}
\title{Recalculate c_top with virtual storage}
\usage{
recalculate_ctop_with_virtualstorage(atm, tlevel, crit_v_top = -0.05)
}
\arguments{
\item{atm}{atmosphere time series data}

\item{tlevel}{tlevel of model output (use \code{kwb.hydrus1d::read_tlevel})}

\item{crit_v_top}{critical v_top rate (default: - 0.05)}
}
\value{
tibble with modified c_top values in order to respect that evaporation
is often larger compared to precipitation
}
\description{
Recalculate c_top with virtual storage
}
