% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_hydrologic_years.R
\name{select_hydrologic_years}
\alias{select_hydrologic_years}
\title{Select hydrologic years}
\usage{
select_hydrologic_years(atm = prepare_atmosphere_data())
}
\arguments{
\item{atm}{atm as retrieved by \code{prepare_atmosphere_data}}
}
\value{
select hydrologic years
}
\description{
Select hydrologic years
}
