% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere_input.R
\name{prepare_atmosphere}
\alias{prepare_atmosphere}
\title{Prepare Atmosphere}
\usage{
prepare_atmosphere(atm, defaults = kwb.hydrus1d::defaults_atmosphere())
}
\arguments{
\item{atm}{atm as retrieved by \code{prepare_atmosphere_data}}

\item{defaults}{defaults for undefined parameters\code{\link[kwb.hydrus1d:defaults_atmosphere]{kwb.hydrus1d::defaults_atmosphere()}}}
}
\value{
tibble with peoered
}
\description{
Prepare Atmosphere
}
\examples{
atm <- prepare_atmosphere_data()
atm_selected <- select_hydrologic_years(atm)
prepare_atmosphere(atm_selected)
}
